+++ community-stabilization-fund-api/package.json
@@ -19,6 +19,7 @@
     "carbon-components": "^11.12.0",
     "carbon-components-react": "^8.12.0",
     "carbon-icons": "^7.0.7",
+    "dayjs": "^1.11.9",
     "next": "12.2.5",
     "react": "18.2.0",
     "react-dom": "18.2.0",
